Add Newton Solver for modes below threshold
It is possible to track single modes as a function of an arbitrary
parameter using a fast Newton solver. This parameter could e.g. be the
pump strength D0 or cavity deformation parameter.

Since there is no Spatial Hole Burning term below the threshold, we
can formulate the Newton problem using complex-valued
matrices/vectors. This is in contrast to the above threshold case, for
which we had to formulate the problem using real-valued
matrices/vectors.