Introduce dof_at_maximum
Instead of having a fixed dof for all the modes of a model laser (see
bcs_norm_constraint that has to be explicitly determined for each mode
l system), it is simpler to automatically determine it for each mode.

In this change we determine the dof of the mode at which the mode has
its highest abs value.