Improve esthzy fig7 tests (support 2 active modes)
In the course of this change it was noticed that the shut-off of the
first and the turn-on of the second mode happens at smaller D0 than in
the published results from the esthzy paper.

It needs to be better studied if there is still a bug in my code.